+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+ - Renamed provider ID from `llamaswap` to `llama-swap`
  - Switch default provider from `@ai-sdk/anthropic` to `@ai-sdk/openai-compatible` with comment
  - Replace all occurences of `:12434` port to `:8080` to signal it's self-hosted